Add peekBackground to HoloPullToRefreshLayout
Used as an optimization for when the user knows that they don't need
to draw a background until peeking under the refreshing view. The
background is only drawn in the non REST state. This saves one level of
overdrawn while just scrolling around.

Combining this with getWindow().setBackgroundDrawable(null) in the
calling app reduces overdraw by 2 total levels.